Move logger back to internal/ since it depends on other internal packages, and don't import it in the api module
