+++ api/go.mod
@@ -4,15 +4,8 @@ go 1.18
 
 require (
 	github.com/PuerkitoBio/rehttp v1.1.0
-	github.com/superfly/flyctl v0.0.350
 	github.com/superfly/graphql v0.2.3
-	golang.org/x/net v0.0.0-20220127200216-cd36cc0744dd
+	golang.org/x/net v0.0.0-20220706163947-c90051bbdb60
 )
 
-require (
-	github.com/alecthomas/chroma v0.10.0 // indirect
-	github.com/blang/semver v3.5.1+incompatible // indirect
-	github.com/dlclark/regexp2 v1.4.0 // indirect
-	github.com/logrusorgru/aurora v2.0.3+incompatible // indirect
-	github.com/pkg/errors v0.9.1 // indirect
-)
+require github.com/pkg/errors v0.9.1 // indirect
