+++ internal/command/auth/auth.go
@@ -19,8 +19,8 @@ import (
 	"github.com/superfly/flyctl/client"
 	"github.com/superfly/flyctl/internal/command"
 	"github.com/superfly/flyctl/internal/config"
+	"github.com/superfly/flyctl/internal/logger"
 	"github.com/superfly/flyctl/internal/state"
-	"github.com/superfly/flyctl/logger"
 )
 
 // New initializes and returns a new apps Command.
