+++ internal/command/command.go
@@ -21,8 +21,8 @@ import (
 	"github.com/superfly/flyctl/internal/buildinfo"
 	"github.com/superfly/flyctl/internal/config"
 	"github.com/superfly/flyctl/internal/env"
+	"github.com/superfly/flyctl/internal/logger"
 	"github.com/superfly/flyctl/internal/update"
-	"github.com/superfly/flyctl/logger"
 
 	"github.com/superfly/flyctl/internal/app"
 	"github.com/superfly/flyctl/internal/cache"
