+++ internal/command/logs/logs.go
@@ -20,8 +20,8 @@ import (
 	"github.com/superfly/flyctl/internal/command"
 	"github.com/superfly/flyctl/internal/config"
 	"github.com/superfly/flyctl/internal/flag"
+	"github.com/superfly/flyctl/internal/logger"
 	"github.com/superfly/flyctl/internal/render"
-	"github.com/superfly/flyctl/logger"
 )
 
 func New() (cmd *cobra.Command) {
